+++ negotiation/src/test/java/it/eng/negotiation/service/ContractNegotiationAPIServiceTest.java
@@ -37,6 +37,7 @@ import it.eng.negotiation.repository.ContractNegotiationRepository;
 import it.eng.negotiation.repository.OfferRepository;
 import it.eng.negotiation.serializer.Serializer;
 import it.eng.tools.client.rest.OkHttpRestClient;
+import it.eng.tools.model.IConstants;
 import it.eng.tools.response.GenericApiResponse;
 import it.eng.tools.util.CredentialUtils;
 
@@ -77,7 +78,8 @@ public class ContractNegotiationAPIServiceTest {
 		
 		service.startNegotiation(MockObjectUtil.FORWARD_TO, Serializer.serializePlainJsonNode(MockObjectUtil.OFFER));
 		
-		verify(contractNegotiationRepository).save(any(ContractNegotiation.class));
+		verify(contractNegotiationRepository).save(argCaptorContractNegotiation.capture());
+		assertEquals(IConstants.ROLE_CONSUMER, argCaptorContractNegotiation.getValue().getRole());
 		verify(offerRepository).save(any(Offer.class));
 	}