+++ CHANGELOG.md
@@ -12,6 +12,16 @@ All notable changes to this project will be documented in this file.
  - Postman collection for testing endpoints	
  - Configured GitHub actions to run tests
 
+## [0.0.1] - 27-08-2024
+
+### Added
+
+ - Added role (consumer or provider) to Contract Negotiation
+ 
+### Removed
+
+ - Removed consumerPid and providerPid from Offer
+
 ## [0.0.1] - 07-08-2024
 
 ### Added
@@ -22,8 +32,8 @@ All notable changes to this project will be documented in this file.
  
 ### Changed
 
- - Agreement service (filter download url) in data transfer module sends request to check if agreement is valid 
-
+ - Agreement service (filter download url) in data transfer module sends request to check if agreement is valid
+ 
 ## [0.0.1] - 06-08-2024
 
 ### Added